fix(meta): fix building wrong package
Prior to this commit, the built package may include `SlowSoftI2CMaster`
instead of `SlowSoftWire`.

It seems like this might be caused by the fact that the source archive
filename is identical between the two libraries, causing the wrong one to
be used.

In this commit, explicitly set the download source filename to include
`SlowSoftWire`.  This fixes the issue.